Blogsite/requirements.txt
Django>=5.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2025.0
django-ckeditor>=6.0.0
django-taggit>=4.0.0
djangorestframework>=3.14.0
Pillow>=10.0.0
